requirements.txt
azure-ai-ml
